requirements.txt
fastapi
uvicorn
openai
langchain
tiktoken
pydantic
python-
google-genai